requirements.txt
alembic==1.11.1
anyio==4.8.0
asyncpg==0.28.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.100.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
pydantic==1.10.21
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
